requirements.txt
# -*- coding: utf-8 -*-
#
# Python requirements for build.py script
#
python-matrix-runner~=1.0
